+++ src/index.ts
@@ -5,7 +5,9 @@ export {
   parseManifest,
   emptyManifest,
 } from "./schema/manifest.js";
+export { manifestJsonSchema } from "./schema/json-schema.js";
 export { analyzeRepository, analyzePath } from "./analyze/index.js";
+export { annotateHtml, annotateFile } from "./analyze/annotate.js";
 export {
   registerManifest,
   getModelContext,